requirements.txt
# 文件处理
python-docx>=1.1.0
PyMuPDF>=1.24.0
lxml>=5.0.0

# AI 接口（按需安装其中一个或全部）
openai>=1.40.0              # DeepSeek（兼容 OpenAI 协议）
google-generativeai>=0.8.0  # Google Gemini
anthropic>=0.40.0           # Anthropic Claude
